Refactor
Selectors from provider to hook.
