+++ src/source/components/common/Navigation/providers/NavigationProvider/NavigationProviderTypes.ts
@@ -10,41 +10,8 @@ export interface NavigationContextStoredValueProps {
 export interface NavigationContextReturnValueProps {
   componentActive: boolean;
   setComponentActive: (componentActive: boolean) => void;
-  closeOpenSiblings: (currentlyFocusedEl: FocusableElement) => void;
-  handleClickAwayClose: (parentEl: HTMLButtonElement | null) => void;
-  getNavObjectByParent: (
-    parentElement: HTMLButtonElement | null,
-  ) => NavigationContextStoredValueProps;
-  getFirstChildElement: (
-    parentEl: HTMLButtonElement | null,
-  ) => FocusableElement;
-  getIsFirstOrLastItem: (focusableEl: FocusableElement) => boolean;
-  getLastChildInRow: (index: number) => FocusableElement;
-  getLastTopElement: (
-    focusableEl: HTMLAnchorElement,
-  ) => FocusableElement | undefined;
-  getNextElement: (
-    parentEl: HTMLButtonElement | null,
-    focusableEl: FocusableElement,
-    currentFocusedList: FocusableElement[],
-    currentKey?: KeyboardEvent["key"],
-  ) => FocusableElement;
-  getPreviousElement: (
-    parentEl: HTMLButtonElement | null,
-    focusableEl: FocusableElement,
-    currentFocusedList: FocusableElement[],
-    isListOpen: boolean,
-    currentKey?: KeyboardEvent["key"],
-  ) => FocusableElement;
   getNavigationArray: () => NavigationContextStoredValueProps[];
-  getSubNavigation: (parentEl) => NavigationContextStoredValueProps[];
-  handleFocusableFocus: (
-    focusableEl: FocusableElement,
-    closeOpenSiblings: (
-      focusableEl: FocusableElement,
-    ) => FocusableElement | undefined,
-  ) => void;
-  // isTopRow: (parentEl: HTMLButtonElement | null) => boolean;
+
   registerNavItem: (
     navList: FocusableElement[],
     parentEl: HTMLButtonElement | null,
